fsck: disable forced full file system scans on version change
f2fs-tools keeps track of the kernel version by saving the content
of /proc/version and forces a full file system scan when this changes
from one boot to another.

Since /proc/version includes a build timestamp, the scan is performed
every time the kernel is re-built.

Enable this check only when it's a release build.

Change-Id: I7da0dc1c89f153fd9de0cc39cdc349679b4215e9
Signed-off-by: Park Ju Hyung <qkrwngud825@gmail.com>