fsck/Makefile.am, mkfs/Makefile.am: fix to `make distcheck` correctly
We add noinst_HEADERS to include header files which are needed to build
but which do not need to be installed on the file system, and include_HEADERS
which are header files that should be installed in $(includedir).
This fixes 'make distcheck`.

Signed-off-by: Anthony G. Basile <blueness@gentoo.org>
[Jaegeuk Kim: remove list.h]
Signed-off-by: Jaegeuk Kim <jaegeuk@kernel.org>

Change-Id: Ic5e2ee949a345b201557752a16bc9d9eca658cf3
(cherry picked from commit 65ce8335cc0d99fe60b7738d606079f826e6406e)